fix: and/or priorities ...

+++ libraries/make_libraries.sh
@@ -21,12 +21,12 @@ die()
 
 BUILDDIR="$SCRIPTDIR/$ARCH/build"
 
-[ ! -d "$BUILDDIR" ] && (
+if [ ! -d "$BUILDDIR" ]
+then
 	mkdir -p "$BUILDDIR" || die "Cannot create $BUILDDIR directory"
-)
+fi
 
 cd "$SCRIPTDIR/$ARCH" || die "Cannot cd to $SCRIPTDIR/$ARCH directory"
-
 #############################################################################
 # OpenSSL
 
@@ -35,24 +35,29 @@ openssl_dir=openssl-$openssl_version
 openssl_tarball=$openssl_dir.tar.gz
 openssl_url="http://www.openssl.org/source/$openssl_tarball"
 
-[ ! -d $openssl_dir ] && (
-	echo "**** Download OpenSSL"
-	wget "$openssl_url" -O $openssl_tarball && \
-	tar xzf $openssl_tarball && \
-	rm $openssl_tarball
-) || die "Couldn't checkout openssl"
+if [ ! -d $openssl_dir ]
+then
+	(
+		echo "**** Download OpenSSL" && \
+		wget "$openssl_url" -O $openssl_tarball && \
+		tar xzf $openssl_tarball && \
+		rm $openssl_tarball
+	) || die "Couldn't checkout openssl"
+fi
 
-[ ! -e build/lib/libssl.a -o ! -e build/lib/libcrypto.a ] && (
-	echo "**** Build OpenSSL"
-	cd $openssl_dir && \
-	./config --prefix="$BUILDDIR"                       \
-	         --openssldir="$BUILDDIR/openssl"           \
-	         no-shared no-asm no-threads zlib           \
-	         -fPIC && \
-	make install && \
-	cd -
-) || die "Couldn't build openssl"
-
+if [ ! -e build/lib/libssl.a -o ! -e build/lib/libcrypto.a ]
+then
+	(
+		echo "**** Build OpenSSL"
+		cd $openssl_dir && \
+		./config --prefix="$BUILDDIR"                       \
+				 --openssldir="$BUILDDIR/openssl"           \
+				 no-shared no-asm no-threads zlib           \
+				 -fPIC && \
+		make install && \
+		cd -
+	) || die "Couldn't build openssl"
+fi
 echo "**** OpenSSL up-to-date"
 
 #############################################################################
@@ -63,40 +68,52 @@ msgpack_dir=msgpack-$msgpack_version
 msgpack_tarball=$msgpack_dir.tar.gz
 msgpack_url="http://msgpack.org/releases/cpp/$msgpack_tarball"
 
-[ ! -d $msgpack_dir ] && (
-	echo "**** Download msgpack"
-	wget "$msgpack_url" -O $msgpack_tarball && \
-	tar xzf $msgpack_tarball && \
-	rm $msgpack_tarball
-) || die "Couldn't checkout msgpack"
+if [ ! -d $msgpack_dir ]
+then
+	(
+		echo "**** Download msgpack" && \
+		wget "$msgpack_url" -O $msgpack_tarball && \
+		tar xzf $msgpack_tarball && \
+		rm $msgpack_tarball
+	) || die "Couldn't checkout msgpack"
+fi
 
-[ ! -e build/lib/libmsgpack.a ] && (
-	echo "**** Build msgpack"
-	cd $msgpack_dir && \
-	./configure --prefix="$BUILDDIR" CFLAGS="${CFLAGS} -fPIC" && \
-	make install && \
-	cd -
-) || die "Couldn't build msgpack"
+if [ ! -e build/lib/libmsgpack.a ]
+then
+	(
+		echo "**** Build msgpack" && \
+		cd $msgpack_dir && \
+		./configure --prefix="$BUILDDIR" CFLAGS="${CFLAGS} -fPIC" && \
+		make install && \
+		cd -
+	) || die "Couldn't build msgpack"
+fi
 
-echo "**** msgpack up-to-date"
+	echo "**** msgpack up-to-date"
 
-#############################################################################
-# jsoncpp
+	#############################################################################
+	# jsoncpp
 
-[ ! -d jsoncpp ] && (
-	echo "**** Download jsoncpp"
-	svn co https://jsoncpp.svn.sourceforge.net/svnroot/jsoncpp/trunk/jsoncpp jsoncpp
-) || die "Couldn't checkout jsoncpp"
+if [ ! -d jsoncpp ]
+then
+	(
+		echo "**** Download jsoncpp" && \
+		svn co https://jsoncpp.svn.sourceforge.net/svnroot/jsoncpp/trunk/jsoncpp jsoncpp
+	) || die "Couldn't checkout jsoncpp"
+fi
 
-[ ! -e build/lib/libjson.a ] && (
-	echo "**** Build jsoncpp"
-	cd jsoncpp && \
-	scons platform=linux-gcc && \
-	cd - && \
-	cp jsoncpp/buildscons/*/src/lib_json/*.a build/lib/libjson.a && \
-	cp jsoncpp/buildscons/*/src/lib_json/*.so build/lib/libjson.so && \
-	cp -r jsoncpp/include/json build/include/jsoncpp
-) || die "Could'nt build jsoncpp"
+if [ ! -e build/lib/libjson.a ]
+then
+	(
+		echo "**** Build jsoncpp" && \
+		cd jsoncpp && \
+		scons platform=linux-gcc && \
+		cd - && \
+		cp jsoncpp/buildscons/*/src/lib_json/*.a build/lib/libjson.a && \
+		cp jsoncpp/buildscons/*/src/lib_json/*.so build/lib/libjson.so && \
+		cp -r jsoncpp/include/json build/include/jsoncpp
+	) || die "Could'nt build jsoncpp"
+fi
 
 echo "**** jsoncpp up-to-date"
 
@@ -108,45 +125,55 @@ curl_dir=curl-$curl_version
 curl_tarball=$curl_dir.tar.gz
 curl_url="http://curl.haxx.se/download/$curl_tarball"
 
-[ ! -d $curl_dir ] && (
-	echo "**** Download curl"
-	wget "$curl_url" -O $curl_tarball && \
-	tar xzf $curl_tarball && \
-	rm $curl_tarball
-) || die "Couldn't checkout curl"
+if [ ! -d $curl_dir ]
+then
+	(
+		echo "**** Download curl" && \
+		wget "$curl_url" -O $curl_tarball && \
+		tar xzf $curl_tarball && \
+		rm $curl_tarball
+	) || die "Couldn't checkout curl"
+fi
 
 
-[ ! -e build/lib/libcurl.a ] && (
-	cd $curl_dir
-	./configure --prefix="$BUILDDIR" --with-ssl="$BUILDDIR/openssl"
-	make install
-	cd -
-) || die "Couldn't build curl"
+if [ ! -e build/lib/libcurl.a ]
+then
+	(
+		cd $curl_dir && \
+		./configure --prefix="$BUILDDIR" --with-ssl="$BUILDDIR/openssl" && \
+		make install
+	) || die "Couldn't build curl"
+fi
 
-echo "**** curl up-to-date"
+	echo "**** curl up-to-date"
 
 
-#############################################################################
-# qjson
+	#############################################################################
+	# qjson
 
-[ ! -d qjson ] && (
-	echo "**** Download qjson"
-	git clone git://gitorious.org/qjson/qjson.git
-) || die "Couldn't checkout qjson"
+if [ ! -d qjson ]
+then
+	(
+		echo "**** Download qjson" && \
+		git clone git://gitorious.org/qjson/qjson.git
+	) || die "Couldn't checkout qjson"
+fi
 
-[ -z "`ls build/lib/libqjson*`" ] && (
-	echo "**** build qjson"
-	cd qjson
-	mkdir build
-	cd build
-	cmake -DCMAKE_BUILD_TYPE=RELEASE -DCMAKE_INSTALL_PREFIX="$BUILDDIR" ..
-	make install
-	cd ../..
-	mkdir -p build/include/QJson
-	cd build/include/QJson
-	echo '#include "../qjson/serializer.h"' > Serializer
-	echo '#include "../qjson/parser.h"' > Parser
-	echo '#include "../qjson/qobjecthelper.h"' > QObjectHelper
-) || die "Couldn't build qjson"
-
+if [ -z "`ls build/lib/libqjson*`" ]
+then
+	(
+		echo "**** build qjson" && \
+		cd qjson && \
+		mkdir build && \
+		cd build && \
+		cmake -DCMAKE_BUILD_TYPE=RELEASE -DCMAKE_INSTALL_PREFIX="$BUILDDIR" .. && \
+		make install && \
+		cd ../.. && \
+		mkdir -p build/include/QJson && \
+		cd build/include/QJson && \
+		echo '#include "../qjson/serializer.h"' > Serializer && \
+		echo '#include "../qjson/parser.h"' > Parser && \
+		echo '#include "../qjson/qobjecthelper.h"' > QObjectHelper
+	) || die "Couldn't build qjson"
+fi
 echo "**** qjson up-to-date"
